first step to interning

+++ rd-cpp/src/rd_core_cpp/src/main/util/gen_util.h
@@ -33,7 +33,7 @@ namespace rd {
 	template<typename T>
 	typename std::enable_if<!std::is_integral<T>::value, size_t>::type
 	contentDeepHashCode(std::vector<T> const &value) noexcept {
-		int result = 1;
+		size_t result = 1;
 		for (auto const &x : value) {
 			result = 31 * result + contentDeepHashCode(x);
 		}
